A lot of changes all around.
The majority of changes revolve around the fact that I've made a new macro to
declare AST class items with the

    class XYZ;
    typedef std::shared_ptr<XYZ> XYZ_p>;

Lines to them. This way, we don't need to declare them in every header file,
and it's all in one central location. As a result, the qualified_id class's
implementations needed to be moved to their own cpp file, which was added as
well.