Make block_statement its own abstract class.
block_statement was already an interface, but it was required derived classes
(i.e. statements) to explicitly call statement_list->whatever(), which should
not be a responsibility of the classes. So now the statement_list method calls
are implemented in the block_statement, and in their tree traversal methods
they call the statement_list methods by using base_t::whatever().