Add function calls and array indexing.
This is a larger change because of the ambiguities between array indexing and function calls. In matlab, they are identical; i.e., a(1) could either be calling some function "a" with argument 1, or it could be indexing a matrix "a" at index 1. These have vastly different R representations, so a "function name visitor" was implemented to collect all of the names of declared functions. This unordered set of names is passed to the codegen visitor. Whenever an array access or funcall is made, it is resolved by checking the set of funcall names.