Change qualified_id in function_name_visitor to expression.
One of the major things that a function rename has to do on occasion is write an entire expression in place of the original function call. Rather than build an entire expression from scratch using the hacky expr field of a qualified_id_item, we're going to begin transitioning to using the expression *containing* a function that needs to be renamed. The expression will be sent as an argument to the mapper function, and modified from there. Thanks to the previous changes where we don't have to go through two extra layers of AST nodes, it'll be much less obnoxious to deal with.