Add build_args helper method
This method is located in qualified_id_item.hpp. This makes sense for its location, because that is the only place where it is used. This function gets used by expression::build_function directly. Because qualified_id_item::build_args uses the array_index_type enum class, a few things in the ast.hpp file have been rearranged and added.